src/model/langchain_service/requirements.txt
python-docx
Flask
flask-cors
openai
python-dotenv
langchain
tiktoken